fix: update types to match the SDK
This updates some fields from int32 to int64 to match the
SDK.  Updating those types exposed an unnecessary use of VLAN
IDs as ASNs in the Metal Gateway tests; those references have
been replaced with a hard-coded ASN of 65100.

+++ internal/projects/bgpenable.go
@@ -47,7 +47,7 @@ func (c *Client) BGPEnable() *cobra.Command {
 func enableBGP(c *Client, config *BgpConfig) error {
 	req := metal.BgpConfigRequestInput{
 		UseCase:        &config.UseCase,
-		Asn:            int32(config.ASN),
+		Asn:            int64(config.ASN),
 		DeploymentType: metal.BgpConfigRequestInputDeploymentType(config.DeploymentType),
 		Md5:            &config.MD5,
 	}
